% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_mdw.r
\docType{data}
\name{Conservation}
\alias{Conservation}
\title{The GPCOG \link{ggplot2} theme}
\format{
An object of class \code{character} of length 1.
}
\usage{
Conservation
}
\description{
Custom color palettes from Maine Design Workshop plans
All credit to Vanessa L. Farr and Kara Wilbur
}
\keyword{datasets}
