% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcog_colors.R
\docType{data}
\name{gpcog_colors}
\alias{gpcog_colors}
\title{The GPCOG \link{ggplot2} theme}
\format{
An object of class \code{list} of length 45.
}
\usage{
gpcog_colors
}
\description{
Custom color palettes from the GPCOG style guide
}
\keyword{datasets}
