% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_gpcog.r
\docType{data}
\name{GPCOG_dark_blue}
\alias{GPCOG_dark_blue}
\title{The GPCOG \link{ggplot2} theme}
\format{
An object of class \code{character} of length 1.
}
\usage{
GPCOG_dark_blue
}
\description{
Custom color palettes from the GPCOG style guide
}
\keyword{datasets}
