% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcog_note.R
\name{gpcog_note}
\alias{gpcog_note}
\title{gpcog_note}
\usage{
gpcog_note(text, size = 8, width = 132, plural = FALSE)
}
\arguments{
\item{text}{character string for a note}

\item{size}{font size for the note}

\item{width}{a number of characters to allow before a character return}

\item{plural}{If \code{TRUE}, will change "Note:" to "Notes:"}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
Add a notes section to the bottom of a plot created with \code{gpcog_plot()}.
}
