% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcog_plot.R
\name{gpcog_plot}
\alias{gpcog_plot}
\title{gpcog_plot}
\usage{
gpcog_plot(..., heights = 1)
}
\arguments{
\item{...}{gpcog plot objects or grobs}

\item{heights}{relative heights of each object in the final plot}
}
\value{
one plot made from many grobs
}
\description{
Combine elements from \code{gpcog_title}, \code{gpcog_subtitle},
\code{gpcog_y_title}, \code{get_legend}, \code{remove_legend},
\code{gpcog_notes}, \code{gpcog_source}, and \code{ggplot2} into
one formatted plot.
}
