% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcog_source.R
\name{gpcog_source}
\alias{gpcog_source}
\title{gpcog_source}
\usage{
gpcog_source(text, size = 8, width = 132, plural = FALSE)
}
\arguments{
\item{text}{character string for a source}

\item{size}{font size for the source}

\item{width}{a number of characters to allow before a character return}

\item{plural}{If \code{TRUE}, will change "Source:" to "Sources:"}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
Add a source section to the bottom of a plot created with \code{gpcog_plot()}.
}
