% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcog_subtitle.R
\name{gpcog_subtitle}
\alias{gpcog_subtitle}
\title{gpcog_subtitle}
\usage{
gpcog_subtitle(string, size = 9.5)
}
\arguments{
\item{string}{character string for a subtitle}

\item{size}{font size for the subtitle}
}
\value{
a grob formatted for a subtitle in a ggplot
}
\description{
Add a subtitle to a plot created with \code{gpcog_plot()}.
}
