% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcog_title.R
\name{gpcog_title}
\alias{gpcog_title}
\title{gpcog_title}
\usage{
gpcog_title(string, size = 12, width = 68)
}
\arguments{
\item{string}{character string for a title}

\item{size}{font size for the title}

\item{width}{a number of characters to allow before a character return}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
Add a title to a plot created with \code{gpcog_plot()}.
}
