% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcogstyle.R
\docType{package}
\name{gpcogstyle}
\alias{gpcogstyle}
\title{GPCOG Style
The GPCOG \link{ggplot2} theme}
\description{
Align "ggplot2" output with the GPCOG style guide, based on the Urban Institute
Data Visualization style guide \url{https://urbaninstitute.github.io/r-at-urban/graphics-guide.html}.
}
\author{
Abe Dailey \email{adailey@gpcog.org}
}
