% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdw_colors.R, R/palette_mdw.r
\docType{data}
\name{mdw_colors}
\alias{mdw_colors}
\title{The GPCOG \link{ggplot2} theme}
\format{
An object of class \code{list} of length 2.

An object of class \code{list} of length 2.
}
\usage{
mdw_colors

mdw_colors
}
\description{
Custom color palettes from Maine Design Workshop Plans
All credit to Vanessa L. Farr and Kara Wilbur
}
\keyword{datasets}
