% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_c.R
\name{scale_color_c}
\alias{scale_color_c}
\title{The GPCOG \link{ggplot2} theme}
\usage{
scale_color_c(name, breaks)
}
\description{
For continuous variables, pass the palette colors from generate_palette() as the values in
the scale_color_manual() function from {ggplot2}
}
