% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_d.R
\name{scale_color_d}
\alias{scale_color_d}
\title{The GPCOG \link{ggplot2} theme}
\usage{
scale_color_d(name, breaks)
}
\description{
For discrete variables, pass the palette colors from generate_palette() as the values in
the scale_colour_manual() function from {ggplot2}
}
