% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_c.R
\name{scale_fill_c}
\alias{scale_fill_c}
\title{The GPCOG \link{ggplot2} theme}
\usage{
scale_fill_c(name, breaks)
}
\description{
For continuous variables, pass the palette colors from generate_palette() as the values in
the scale_fill_manual() function from {ggplot2}
}
