% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gpcog_print.R
\name{theme_gpcog_print}
\alias{theme_gpcog_print}
\title{The GPCOG \link{ggplot2} theme}
\usage{
theme_gpcog_print(
  base_size = 9,
  base_family = "Arial",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{base_line_size, base_rect_size}{base line and rectangle sizes}
}
\description{
\code{theme_gpcog} provides a \link{ggplot2} theme formatted according to the
Urban Institute style guide for web, with sensible defaults.
}
