% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undo_gpcog_defaults.R
\name{undo_gpcog_defaults}
\alias{undo_gpcog_defaults}
\title{The GPCOG \link{ggplot2} theme}
\usage{
undo_gpcog_defaults()
}
\description{
Resets all the \link{ggplot2} theme to its default settings
}
